move loaders to scrapy.loader
